ISSUE-02: moved catalog.py to package pythonsupermarket. adapted dependent files that use it.
